+++ src/components/NetworkList.js
@@ -1,53 +1,51 @@
 // Copyright (c) 2019 SafetyCulture Pty Ltd. All Rights Reserved.
 
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import AutoSizer from 'react-virtualized-auto-sizer';
-import { FixedSizeList as List } from 'react-window';
-import NetworkListRow from './NetworkListRow';
+import React from "react";
+import { useSelector } from "react-redux";
+import AutoSizer from "react-virtualized-auto-sizer";
+import { FixedSizeList as List } from "react-window";
+import NetworkListRow from "./NetworkListRow";
 
-import './NetworkList.css';
+import "./NetworkList.css";
+import { selectFilteredLogs } from "../state/network";
 
-class NetworkList extends Component {
-  render() {
-    const { network } = this.props;
-    return (
-      <div className="widget vbox network-list">
-        <div className="widget vbox">
-          <div className="data-grid">
-            <div className="header-container">
-              <table className="header">
-                <tbody>
-                  <tr>
-                    <th>
-                      <div>Name</div>
-                    </th>
-                  </tr>
-                </tbody>
-              </table>
-            </div>
-            <div className="data-container">
-              <AutoSizer disableWidth>
-                {({ height }) => (
-                  <List
-                    className="data"
-                    itemCount={network.log.length}
-                    height={height}
-                    itemSize={21}
-                    itemData={network.log}
-                    overscanCount={50}
-                  >
-                    {NetworkListRow}
-                  </List>
-                )}
-              </AutoSizer>
-            </div>
+const NetworkList = () => {
+  const log = useSelector(selectFilteredLogs);
+  return (
+    <div className="widget vbox network-list">
+      <div className="widget vbox">
+        <div className="data-grid">
+          <div className="header-container">
+            <table className="header">
+              <tbody>
+                <tr>
+                  <th>
+                    <div>Name</div>
+                  </th>
+                </tr>
+              </tbody>
+            </table>
+          </div>
+          <div className="data-container">
+            <AutoSizer disableWidth>
+              {({ height }) => (
+                <List
+                  className="data"
+                  itemCount={log.length}
+                  height={height}
+                  itemSize={21}
+                  itemData={log}
+                  overscanCount={50}
+                >
+                  {NetworkListRow}
+                </List>
+              )}
+            </AutoSizer>
           </div>
         </div>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
-const mapStateToProps = state => ({ network: state.network })
-export default connect(mapStateToProps)(NetworkList)
+export default NetworkList;