+++ src/icons/Clear.js
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { PureComponent } from "react";
 
 class Clear extends PureComponent {
   render() {