+++ src/icons/Down.js
@@ -1,10 +1,13 @@
-import React, { PureComponent } from 'react';
+import React, { PureComponent } from "react";
 
 class Down extends PureComponent {
   render() {
     return (
       <svg width="14" viewBox="0 0 14 14">
-        <path d="M7.935,10.205L6.242,11.899L12.517,12.517L11.899,6.242L10.205,7.935L3.752,1.483L1.483,3.752L7.935,10.205Z" />
+        <path
+          className="arrow-down"
+          d="M7.935,10.205L6.242,11.899L12.517,12.517L11.899,6.242L10.205,7.935L3.752,1.483L1.483,3.752L7.935,10.205Z"
+        />
       </svg>
     );
   }