+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 2.1.0 - 2023.10.03
+
+- feat: add stop/resume button.
+- feat: add service name if the method is `url.XxxxService/path` on the list.
+- fix: fix filters crashing.
+- chore: remove `fuse.js` library.
+
 ## 2.0.1 - 2023.07.28
 
 - Remove log.